hypermap-cacher: refactor into multiple files
